requirements.txt
beautifulsoup4<5
nats-py[nkeys]<3
requests<3
